Split utils module into separate files
