+++ cmdstanpy/install_cxx_toolchain.py
@@ -27,7 +27,7 @@ from cmdstanpy import _DOT_CMDSTAN
 from cmdstanpy.utils import pushd, validate_dir, wrap_url_progress_hook
 
 EXTENSION = '.exe' if platform.system() == 'Windows' else ''
-IS_64BITS = sys.maxsize > 2**32
+IS_64BITS = sys.maxsize > 2 ** 32
 
 
 def usage() -> None:
